Moving emotion circle cell into its own file.
